Refactor chat block into own app
