+++ pydatalab/pydatalab/blocks/blocks.py
@@ -2,13 +2,11 @@ import base64
 import io
 import os
 import random
-import textwrap
 import zipfile
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple
 
 import bokeh.embed
 import numpy as np
-import openai
 import pandas as pd
 from bokeh.models import Text
 from bokeh.plotting import figure
@@ -409,99 +407,3 @@ class XRDBlock(DataBlock):
             )
 
             self.data["bokeh_plot_data"] = bokeh.embed.json_item(p, theme=mytheme)
-
-
-class ChatBlock(DataBlock):
-
-    blocktype = "chat"
-    description = "LLM Chat Block with contextual data (powered by GPT-3.5-turbo)"
-    accepted_file_extensions = []
-
-    defaults = {
-        "system_prompt": """You are C.L.A.I.R.E., an LLM interface for cybernetically enhanced materials chemistry in the group of Professor Dame Clare Grey FRS in the Department of Chemistry at the University of Cambridge.
-
-You have been embedded within the group's data management solution, datalab, where you have access to recorded data on samples, electrochemical cells and characterisation experiments (XRD, NMR etc.). You will be asked questions about the data and you are tasked with providing helpful responses to questions that researcher's may ask about their data. You are currently only available in a trial mode but one day we would like you to learn from the group wiki at https://wikis.ch.cam.ac.uk/grey/wiki/index.php/Main_Page to allow you to provide even more helpful answers!
-
-We look forward to working with you!""",
-    }
-    openai.api_key = os.environ.get("OPENAI_API_KEY")
-
-    @property
-    def plot_functions(self):
-        return (self.render,)
-
-    def render(self):
-
-        item_info = flask_mongo.db.items.find_one(
-            {"item_id": self.data["item_id"]},
-            projection={"blocks_obj": 0},
-        )
-
-        if not self.data.get("messages"):
-            self.data["messages"] = [
-                {
-                    "role": "system",
-                    "content": self.defaults["system_prompt"],
-                },
-                {
-                    "role": "user",
-                    "content": f"Below is the JSON API response for this sample, please start with a brief description of the experiment after a friendly welcome message of your choice. Please only reference the data in the JSON response, do not reference the data in the wiki or any other source. JSON data: {item_info}",
-                },
-            ]
-
-        if self.data.get("prompt"):
-            self.data["messages"].append(
-                {
-                    "role": "user",
-                    "content": self.data["prompt"],
-                }
-            )
-            self.data["prompt"] = None
-
-        responses = openai.ChatCompletion.create(
-            model="gpt-3.5-turbo",
-            messages=self.data["messages"],
-            temperature=0.9,
-            max_tokens=1024,
-            stop=["\n"],
-        )
-
-        self.data["messages"].append(responses["choices"][0].message)
-
-        p = figure(
-            sizing_mode="scale_width",
-            aspect_ratio=1.5,
-            tools=[],
-        )
-        p.axis.visible = False
-        p.xgrid.grid_line_color = None
-        p.ygrid.grid_line_color = None
-
-        for ind, message in enumerate(self.data["messages"][2:]):
-            LOGGER.debug(message)
-            try:
-                role = message.role
-                content = message.content
-            except:
-                role = message["role"]
-                content = message["content"]
-
-            if role == "user":
-                color = "blue"
-                x = 0
-            else:
-                color = "green"
-                x = 0
-            y = -ind
-
-            text = [line for line in textwrap.wrap(content, 80)]
-
-            bubble = Text(x=x, y=y, text=["\n".join(text)], text_color=color)
-            p.add_glyph(bubble)
-
-        from bokeh.models import Range1d
-
-        p.x_range = Range1d(-1, 40)
-        p.y_range = Range1d(-5, 5)
-
-        self.data["bokeh_plot_data"] = bokeh.embed.json_item(p)
