+++ webapp/src/components/datablocks/ChatBlock.vue
@@ -2,22 +2,15 @@
   <!-- think about elegant two-way binding to DataBlockBase... or, just pass all the block data into
 DataBlockBase as a prop, and save from within DataBlockBase  -->
   <DataBlockBase :item_id="item_id" :block_id="block_id">
-    <div class="form-row col-md-4 col-lg-4 mt-2 mb-2 pl-1">
-      <div class="input-group form-inline">
-        <label class="mr-2"><b>Prompt</b></label>
-        <input
-          type="text"
-          class="form-control"
-          v-model="prompt"
-          @keydown.enter="updateBlock()"
-          @blur="updateBlock()"
-        />
-      </div>
-    </div>
-
     <div class="row">
       <div id="bokehPlotContainer" class="col-xl-9 col-lg-10 col-md-11 mx-auto">
         <BokehPlot :bokehPlotData="bokehPlotData" />
+      </div>
+    </div>
+    <div class="form-row col-md-10 col-lg-4 mt-2 mb-2 pl-1">
+      <div class="input-group form-inline">
+        <label class="mr-2"><b>Prompt</b></label>
+        <input type="text" class="form-control" v-model="prompt" @keydown.enter="updateBlock()" />
       </div>
     </div>
   </DataBlockBase>
